# lib/api-zod/src/generated/types/eventFeedbackReceipt.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Squabblemon player and game-loop API
 * OpenAPI spec version: 0.2.0
 */
import type { EventFeedbackPost } from './eventFeedbackPost';

export interface EventFeedbackReceipt {
  post: EventFeedbackPost;
  /** The saved post ID; unchanged on retries. */
  receiptId: string;
  replayed: boolean;
}