# lib/api-zod/src/generated/types/socialMatchOpponent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Squabblemon player and game-loop API
 * OpenAPI spec version: 0.2.0
 */
import type { SocialLookup } from './socialLookup';

export interface SocialMatchOpponent {
  opponent: SocialLookup | null;
}