# lib/api-zod/src/generated/types/socialPlayer.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Squabblemon player and game-loop API
 * OpenAPI spec version: 0.2.0
 */

export interface SocialPlayer {
  username: string;
  lastActiveAt: string | null;
  friendCode: string;
  displayName: string;
  avatarKey: string;
}